Create symbolic link only if the target exists

+++ weechat/download/models.py
@@ -239,10 +239,11 @@ def handler_package_saved(sender, **kwargs):
 
 def make_symlink(link_name, target):
     """Create a symbolic link (overwrite link_name if existing)."""
-    print(f'Making symlink: {link_name} -> {target}')
-    tmp_name = f'{link_name}.__tmp__'
-    os.symlink(target, tmp_name)
-    os.rename(tmp_name, link_name)
+    if os.path.exists(target):
+        print(f'Making symlink: {link_name} -> {target}')
+        tmp_name = f'{link_name}.__tmp__'
+        os.symlink(target, tmp_name)
+        os.rename(tmp_name, link_name)
 
 
 def set_stable_version(project, version):
